requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
dj-database-url==1.0.0
Django==4.1.3
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-daraja==1.2.0
django-heroku==0.3.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
psycopg2-binary==2.9.5
pycparser==2.21
python-decouple==3.6
python-dotenv==0.21.0
python-http-client==3.3.7
pytz==2022.6
requests==2.28.1
sendgrid==6.9.7
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
urllib3==1.26.13
whitenoise==6.2.0
